photo upload changes
